backend-common: refactor reader types
